sparate category list form
