+++ src/components/Category/Form/type.ts
@@ -1,0 +1,3 @@
+export type FormInputType = {
+  name: string;
+};
