+++ src/components/Category/List/store.ts
@@ -1,0 +1,64 @@
+import { ChangeEvent, useState } from 'react';
+import { useSnackbar } from 'notistack';
+import { useTranslation } from 'react-i18next';
+import { useCategory } from 'src/hooks/useCategory';
+
+export const useList = () => {
+  const { t }: { t: any } = useTranslation();
+  const { enqueueSnackbar } = useSnackbar();
+  const { getCategories, deleteCategory, categories, totalCategoryCount } =
+    useCategory();
+
+  const [deleteId, setDeletedId] = useState<number>(0);
+  const [page, setPage] = useState<number>(0);
+  const [limit, setLimit] = useState<number>(10);
+  const [query, setQuery] = useState<string>('');
+
+  const handleQueryChange = (event: ChangeEvent<HTMLInputElement>): void => {
+    event.persist();
+    setQuery(event.target.value);
+  };
+
+  const handlePageChange = (_event: any, newPage: number): void => {
+    setPage(newPage);
+  };
+  const handleLimitChange = (event: ChangeEvent<HTMLInputElement>): void => {
+    setLimit(parseInt(event.target.value));
+  };
+
+  const [openConfirmDelete, setOpenConfirmDelete] = useState(false);
+
+  const handleConfirmDelete = () => {
+    setOpenConfirmDelete(true);
+  };
+
+  const closeConfirmDelete = () => {
+    setOpenConfirmDelete(false);
+  };
+
+  const handleDeleteCompleted = () => {
+    deleteCategory(deleteId, () => {
+      setOpenConfirmDelete(false);
+    });
+  };
+
+  const store = {
+    categories,
+    totalCategoryCount,
+    page,
+    limit,
+    query,
+    openConfirmDelete,
+
+    setDeletedId,
+    getCategories,
+    handleQueryChange,
+    handlePageChange,
+    handleLimitChange,
+    handleConfirmDelete,
+    closeConfirmDelete,
+    handleDeleteCompleted
+  };
+
+  return store;
+};
